Abstracting some of the kcpUtils into a more general cardUtils
