+++ packages/playground/tests/e2e/constants.js
@@ -1,5 +1,5 @@
 export const KOREAN_TEST_CARD = '9490220006611406'; // 9490220006611406 works against Test. For localhost:8080 use: 5067589608564358 + hack in triggerBinLookup
-export const NON_KOREAN_TEST_CARD = '5500000000000004';
+export const REGULAR_TEST_CARD = '5500000000000004';
 export const DUAL_BRANDED_CARD = '4035501000000008'; // dual branded visa & cartebancaire
 
 export const TEST_DATE_VALUE = '10/20';
